Election module/move helper (#799)
* Moved helper

* Remove old helper

* Fix
